Blueprint page refactoring
