+++ frontend-nextjs/src/app/blueprint/page.tsx
@@ -2,7 +2,7 @@
 
 import dynamic from "next/dynamic";
 import { useEffect, useState } from "react";
-import { fetchBlueprintData } from "@/app/blueprinnt/action";
+import { fetchBlueprintData } from "@/app/blueprint/action";
 import AutomationPanel from "@/components/blueprint-page/AutomationPanel";
 import { Blueprint } from "@/type/blueprint";
 
@@ -32,7 +32,7 @@ export default function BlueprintPage() {
                 setCurrentBlueprint(latestBlueprint);
 
                 // Fetch automation list
-                const automationResponse = await fetch("http://localhost:8000/automations");
+                const automationResponse = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_DOMAIN}/automations`);
                 const automationData = await automationResponse.json();
                 setAutomations(automationData);
